fix(messaging): address code review on callback response routing
- pass callback CorrelationId/CorrelationSequence via typed PublishOptions
  properties instead of reserved Headers keys (publish no longer throws)
- add ConsumeContext.SetNextCallback for multi-hop chaining (-> PublishOptions.CallbackName)
- include MessageType in PublishOptions Equals/GetHashCode
- constrain SetResponse<TResponse> to class; internal sealed ConsumerExecutedResult/ISubscribeInvoker
- assert Headers.Type in queue round-trip; extract callback test fixtures
- add copyright headers to demo files

Integration suite: 24/24 green (NatsPostgreSql), callback round-trip verified.
