+++ demo/Headless.Messaging.Console.Demo/EventConsumer.cs
@@ -1,3 +1,5 @@
+// Copyright (c) Mahmoud Shaheen. All rights reserved.
+
 using Headless.Messaging;
 
 namespace Demo;
