+++ src/Headless.Messaging.Core/Internal/ConsumerExecutedResult.cs
@@ -2,12 +2,13 @@
 
 namespace Headless.Messaging.Internal;
 
-public class ConsumerExecutedResult(
+internal sealed class ConsumerExecutedResult(
     object? result,
     Type? resultType,
     string msgId,
     string? callbackName,
-    IDictionary<string, string?>? callbackHeader
+    IDictionary<string, string?>? callbackHeader,
+    string? nextCallbackName = null
 )
 {
     public object? Result { get; set; } = result;
@@ -16,7 +17,15 @@ public class ConsumerExecutedResult(
 
     public string MessageId { get; set; } = msgId;
 
+    /// <summary>The message name the captured response is published to (the current message's callback destination).</summary>
     public string? CallbackName { get; set; } = callbackName;
 
     public IDictionary<string, string?>? CallbackHeader { get; set; } = callbackHeader;
+
+    /// <summary>
+    /// The callback name to stamp on the published response so the next hop in a callback chain can
+    /// react to it. Captured from <c>ConsumeContext.SetNextCallback</c>; maps to
+    /// <c>PublishOptions.CallbackName</c> on the response publish.
+    /// </summary>
+    public string? NextCallbackName { get; set; } = nextCallbackName;
 }
