+++ src/Headless.Messaging.Core/Internal/ISubscribeExecutor.cs
@@ -457,14 +457,13 @@ internal sealed class SubscribeExecutor(
 
             if (!string.IsNullOrEmpty(ret.CallbackName))
             {
-                ret.CallbackHeader ??= new Dictionary<string, string?>(StringComparer.Ordinal);
-                ret.CallbackHeader[Headers.CorrelationId] = message.Origin.GetId();
-                ret.CallbackHeader[Headers.CorrelationSequence] = (
-                    message.Origin.GetCorrelationSequence() + 1
-                ).ToString(CultureInfo.InvariantCulture);
-
+                // TraceParent is NOT a reserved header, so it rides in PublishOptions.Headers alongside
+                // any user-supplied AddResponseHeader keys. The correlation identifiers ARE reserved
+                // (MessagePublishRequestFactory rejects them as custom headers), so they must flow through
+                // the typed PublishOptions surface instead of CallbackHeader.
                 if (message.Origin.Headers.TryGetValue(Headers.TraceParent, out var traceParent))
                 {
+                    ret.CallbackHeader ??= new Dictionary<string, string?>(StringComparer.Ordinal);
                     ret.CallbackHeader[Headers.TraceParent] = traceParent;
                 }
 
@@ -477,6 +476,11 @@ internal sealed class SubscribeExecutor(
                             MessageName = ret.CallbackName,
                             Headers = ret.CallbackHeader,
                             MessageType = ret.ResultType,
+                            CorrelationId = message.Origin.GetId(),
+                            CorrelationSequence = message.Origin.GetCorrelationSequence() + 1,
+                            // Chain the next hop: the published response carries this callback name so its
+                            // consumer can react and publish a further response.
+                            CallbackName = ret.NextCallbackName,
                         },
                         cancellationToken
                     )
